backend/requirements.txt
fastapi
transformers
pydantic
uvicorn
openai
python-multipart
bcrypt
pymongo
python-jose
python-docx
itsdangerous
authlib
httpx
Pillow
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
aiofiles
aiohttp
boto3
roboflow